% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{GGIR_from_csv}
\alias{GGIR_from_csv}
\title{Apply GGIR to Down-Sampled Accelerometer Files (.csv format)}
\usage{
GGIR_from_csv(
  dsdir = c(),
  alloutdir = c(),
  outputdir = c(),
  rmc.col.acc = c(2:4),
  rmc.col.time = 1,
  tz = "UTC"
)
}
\arguments{
\item{dsdir}{Directory for down-sampled files}

\item{alloutdir}{General output directory, default created if not specified}

\item{outputdir}{Directory for GGIR output}

\item{rmc.col.acc}{Columns of accelerometer data in down-sampled files}

\item{rmc.col.time}{Column of timestamps in down-sampled files}

\item{tz}{Timezone (use OlsonNames() for a list of accepted timezone names}
}
\description{
Specifies parameters and implements GGIR (Migueles et. al., 2019) across all .csv accelerometer files (frequency = 1Hz)
in 'acceldir', extracting sleep-wake predictions and sustained inactivity bouts.

Minimum required inputs: 'acceldir'
}
\examples{
GGIR_from_csv(dsdir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/106_output/ds_output")
GGIR_from_csv(dsdir = "~/R/Biobank/Light_Data/265_evenfreq_output/ds_output")

}
