% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{rollingWindowInd}
\alias{rollingWindowInd}
\title{Get Rolling Window Indices}
\usage{
rollingWindowInd(t = c(), window = c(), step = c())
}
\arguments{
\item{t}{Time series vector}

\item{window}{Rolling window size}

\item{step}{Step size for sliding window along time series}
}
\description{
Fits rolling windows to time series and outputs start/end indices of each window.

Note: keep time units consistent between t, window and step. All inputs required.
}
\examples{
rollingWindowInd <- function(t=1:10000, window=3600, step=300)
}
