% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{SRI_from_accel_csv}
\alias{SRI_from_accel_csv}
\title{Calculate Sleep Regularity Index (SRI) from .csv Accelerometer Data}
\usage{
SRI_from_accel_csv(
  acceldir = c(),
  col.timestamp = c(),
  col.accel = c(),
  alloutdir = c(),
  dsdir = c(),
  nwdir = c(),
  outputdir = c(),
  rmc.col.time = 1,
  rmc.col.acc = c(2:4),
  sdThres = 0.12753,
  rngThres = 0.4905,
  tz = "UTC",
  use.naps = TRUE,
  use.WASO = TRUE,
  use.miscal = TRUE,
  use.GGIRnonwear = TRUE,
  use.customnonwear = TRUE,
  nonWearInGGIRsleep = FALSE,
  wr.SWV = TRUE,
  wr.raster = TRUE,
  minSRIdays = 5,
  ...
)
}
\arguments{
\item{acceldir}{Directory containing raw .csv accelerometer files}

\item{col.timestamp}{Column of raw .csv files containing timestamp}

\item{col.accel}{Columns of raw .csv files containing x-y-z accelerometer data e.g., c(1:3)}

\item{alloutdir}{General output directory, default created if not specified}

\item{dsdir}{Directory for down-sampled files}

\item{nwdir}{Directory of non-wear data - specify if using custom (i.e., non-GGIR) non-wear detection. GGIR non-wear data is used preferentially, if available}

\item{outputdir}{Directory of GGIR output}

\item{rmc.col.time}{Column of timestamps in down-sampled files}

\item{rmc.col.acc}{Columns of accelerometer data in down-sampled files}

\item{sdThres}{Standard deviation threshold for non-wear classification, applied per window}

\item{rngThres}{Range threshold for non-wear classification, applied per window}

\item{tz}{Timezone (use OlsonNames() for a list of accepted timezone names)}

\item{use.naps}{Specify whether 'naps' are included in SRI calculation}

\item{use.WASO}{Specify whether 'wake after sleep onset' (WASO) periods are included in SRI calculation}

\item{use.miscal}{Specify whether to filter out nights of 'miscalculated' sleep onset/offset timing}

\item{use.GGIRnonwear}{Specify whether to use GGIR's inbuilt non-wear detection}

\item{use.customnonwear}{Specify whether to use custom (i.e., non-GGIR) non-wear detection (based on van Hees et. al., 2011)}

\item{nonWearInGGIRsleep}{Specify whether non-wear periods within GGIR's 'sleep windows' are to be included as non-wear}

\item{wr.SWV}{Specify whether Sleep-Wake Vectors (SWV) are output to file}

\item{wr.raster}{Specify whether sleep-wake raster plots are output to file}

\item{minSRIdays}{Minimum number of days of overlapping data to calculate valid SRI scores}

\item{...}{}
}
\description{
A wrapper that allows for direct calculation of Sleep Regularity Index (SRI) scores from accelerometer data (.csv format).

Runs four functions by default: (a) down-sampling data [ds_accel_csv], (b) extraction of non-wear data [nonwear_detect],
(c) predicting sleep-wake timing using GGIR [GGIR_from_csv], and (d) calculating SRI after accounting for
fragmented sleep patterns and naps [SRI_from_GGIR].

Minimum required inputs: 'acceldir', 'col.timestamp', 'col.accel'.
}
\examples{
SRI_from_accel_csv(acceldir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/106_output/ds_output", col.timestamp = 5, col.accel = c(1:3))
SRI_from_accel_csv(acceldir = "~/R/Biobank/GGIR_Version_Testing/105_evensample", col.timestamp = 5, col.accel = c(1:3))
SRI_from_accel_csv(acceldir = "~/R/Biobank/Light_Data/127_evenfreq", col.timestamp = 5, col.accel = c(1:3))
}
