% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{SRI_from_binary}
\alias{SRI_from_binary}
\title{Calculate Sleep Regularity Index (SRI) from Binary Sleep-Wake Data}
\usage{
SRI_from_binary(
  binarydir = c(),
  tz = "UTC",
  alloutdir = c(),
  col.trans = 1,
  col.timestamp = 2,
  overwr = FALSE,
  wr.raster = TRUE,
  minSRIdays = 5,
  exclNAhrs = 6
)
}
\arguments{
\item{binarydir}{Directory containing sleep diary data}

\item{tz}{Timezone (use OlsonNames() for a list of accepted timezone names)}

\item{alloutdir}{General output directory, default created if not specified}

\item{col.trans}{Column of sleep-wake transition data}

\item{col.timestamp}{Column of timestamps}

\item{overwr}{Specify whether to overwrite previous SRI data}

\item{wr.raster}{Specify whether to output sleep-wake raster plots}

\item{minSRIdays}{Minimum number of days of overlapping data to calculate valid SRI scores}

\item{exclNAhrs}{Hours of NA per day above which entire day will be excluded (12noon-12noon)}
}
\description{
Calculates SRI from a time series of binary sleep-wake summary (SWS) data in .csv format. Column 1 contains values 1=sleep,
0=wake, NA=NA, end=recording end. Column 2 contains UNIX timestamps (origin=1970-01-01) identifying the start of each associated column 1 value.
Final timestamp specifies recording end.

Minimum required input: 'binarydir'. Specify 'tz' if required (default = "UTC").
}
\examples{
SRI_from_binary(binarydir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/diarydata", tz="Portugal")
SRI_from_binary(binarydir = "~/R/Biobank/Light_Data/127_evenfreq_output/SWS_output", tz="Australia/Melbourne")
SRI_from_binary(binarydir = "~/R/Biobank/Light_Data/127_evenfreq_output/SWS_output_errors")
}
