% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRI_from_diaries.R
\name{SRI_from_diaries}
\alias{SRI_from_diaries}
\title{Calculate Sleep Regularity Index (SRI) from Sleep Diary Data}
\usage{
SRI_from_diaries(
  diarydir = c(),
  alloutdir = c(),
  col.trans = 1,
  col.timestamp = 2,
  overwr = FALSE,
  wr.raster = TRUE
)
}
\arguments{
\item{diarydir}{Directory containing sleep diary data}

\item{alloutdir}{General output directory, default created if not specified}

\item{col.trans}{Column of sleep-wake transition data}

\item{col.timestamp}{Column of timestamp transition data}

\item{overwr}{Specify whether to overwrite previous SRI data}

\item{wr.raster}{Specify whether to output sleep-wake raster plots}
}
\value{

}
\description{
Calculates SRI from sleep-wake data in .csv format, where column 1 contains all sleep-wake transitions (
0 = wake onset, 1 = sleep onset, NA = onset of missing data), column 2 contains UNIX timestamps (seconds, origin = 01/01/1970)
}
\examples{
SRI_from_diaries(diarydir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/diarydata")
}
