% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{SRI_from_GGIR}
\alias{SRI_from_GGIR}
\title{Calculate Sleep Regularity Index (SRI) from GGIR Output}
\usage{
SRI_from_GGIR(
  outputdir = c(),
  alloutdir = c(),
  nwdir = c(),
  use.naps = TRUE,
  use.WASO = TRUE,
  use.miscal = TRUE,
  use.GGIRnonwear = TRUE,
  use.customnonwear = TRUE,
  nonWearInGGIRsleep = FALSE,
  wr.SWV = TRUE,
  wr.raster = TRUE,
  minSRIdays = 5,
  SIdef = "T5A5",
  GGIR_nonwear_detection_stage = "final"
)
}
\arguments{
\item{outputdir}{Directory of GGIR output}

\item{alloutdir}{General output directory, default created if not specified}

\item{nwdir}{Directory of non-wear data - specify if using custom (i.e., non-GGIR) non-wear detection. GGIR non-wear data is used preferentially, if available}

\item{use.naps}{Specify whether 'naps' are included in SRI calculation}

\item{use.WASO}{Specify whether 'wake after sleep onset' (WASO) periods are included in SRI calculation}

\item{use.miscal}{Specify whether to filter out nights of data likely miscalculated by GGIR}

\item{use.GGIRnonwear}{Specify whether to use GGIR's inbuilt non-wear detection}

\item{use.customnonwear}{Specify whether to use custom (i.e., non-GGIR) non-wear detection (based on van Hees et. al., 2011)}

\item{nonWearInGGIRsleep}{Specify whether non-wear periods within GGIR's 'sleep windows' are to be included as non-wear}

\item{wr.SWV}{Specify whether Sleep-Wake Vectors (SWV) are output to file}

\item{wr.raster}{Specify whether sleep-wake raster plots are output to file}

\item{minSRIdays}{Minimum number of days of overlapping data to calculate valid SRI scores}

\item{SIdef}{Sustained inactivity definition (e.g., T5A5)}
}
\description{
Uses sleep windows and sustained inactivity bouts from GGIR output to calculate Sleep Regularity Index scores. Accounts for
naps and fragmented sleep by identifying periods of 'wake' during GGIR-defined sleep windows and periods of 'napping' outside
GGIR-defined sleep windows. Uses sustained inactivity bouts to exclude days where sleep onset and offset times are likely
miscalculated. Runs across all "output_xxx" directories within 'outputdir', accounting for both multi-file and single-file GGIR output
structures.

Additional outputs: sleep-wake raster plots, sleep variables, summary of miscalculated nights, sleep-wake vectors.

Minimum required inputs: 'outputdir', 'nwdir' (if use.customnonwear = TRUE (default))
}
\examples{
SRI_from_GGIR(outputdir = "~/R/Biobank/Light_Data/127_output/GGIR_output")
SRI_from_GGIR(outputdir = "~/R/Biobank/Light_Data/127_evenfreq_output/GGIR_output",
nwdir = "~/R/Biobank/Light_Data/127_evenfreq_output/nw_output")
}
