% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRI_from_accel_csv_v3b.R
\name{ds_accel_csv}
\alias{ds_accel_csv}
\title{Down-Sample Accelerometer Files}
\usage{
ds_accel_csv(
  acceldir = c(),
  alloutdir = c(),
  dsdir = c(),
  col.timestamp = c(),
  col.accel = c()
)
}
\arguments{
\item{acceldir}{Directory containing raw .csv accelerometer files}

\item{alloutdir}{General output directory}

\item{dsdir}{Directory for down-sampled files}

\item{col.timestamp}{Column of raw .csv files containing timestamp}

\item{col.accel}{Columns of raw .csv files containing x-y-z accelerometer data}
}
\value{

}
\description{
Down-samples .csv files to 1Hz, creating required input format and increasing speed.
}
\examples{
ds_accel_csv(acceldir="C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/106", col.timestamp = 1, col.accel = c(2,3,4))

}
