% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{nonwear_detect}
\alias{nonwear_detect}
\title{Non-wear Detection}
\usage{
nonwear_detect(
  dsdir = c(),
  nwdir = c(),
  rmc.col.time = 1,
  rmc.col.acc = c(2, 3, 4),
  sdThres = 0.12753,
  rngThres = 0.4905
)
}
\arguments{
\item{dsdir}{Directory for down-sampled files}

\item{nwdir}{Directory for non-wear data, created under parent output directory if not specified}

\item{rmc.col.time}{Column of down-sampled .csv files containing timestamp}

\item{rmc.col.acc}{Columns of down-sampled .csv files containing x-y-z accelerometer data e.g., c(1,2,3)}

\item{sdThres}{Standard deviation threshold for non-wear classification, applied per window}

\item{rngThres}{Range threshold for non-wear classification, applied per window}
}
\value{

}
\description{
Evaluates epoch-by-epoch non-wear status of accelerometer devices using method described in van Hees et. al. (2011). 15 minute epochs are
evaluated based on surrounding 60 minute windows (centered at 15min), where standard deviation < 13mg and range < 50mg in at least two accelerometer axes
is required for non-wear classification.

Minimum required inputs: 'dsdir', 'rmc.col.time', 'rmc.col.acc'
}
\examples{
nonwear_detect(dsdir = "~/R/Biobank/Light_Data/265_evenfreq_output/ds_output", rmc.col.time = 1, rmc.col.acc = c(2,3,4))
nonwear_detect(dsdir = "~/R/Biobank/Light_Data/127_evenfreq", rmc.col.time = 5, rmc.col.acc = c(1,2,3))
}
