% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{raster_from_SWS}
\alias{raster_from_SWS}
\title{Raster Plot from Sleep-Wake Summary Data}
\usage{
raster_from_SWS(SWS = c(), rasdir = c(), pptName = c(), tz = "UTC")
}
\arguments{
\item{SWS}{Sleep-wake summary data (2-column d.f.). Column 1 contains values 1=sleep, 0=wake, NA=NA, end=recording end.
Column 2 contains UNIX timestamps (origin=1970-01-01) identifying the start of each associated column 1 value.
Final timestamp specifies recording end.}

\item{rasdir}{Raster output directory, default created if not specified}

\item{pptName}{Name of participant or file}

\item{tz}{Timezone (use OlsonNames() for a list of accepted timezone names)}
}
\value{

}
\description{
Extract raster plots from a binary sleep-wake time series (summarized as a reduced-form data frame)
}
\examples{
raster_from_SWS(SWS = SWS, rasdir = "~/R/Biobank/SleepRegPackage/diarydata_output/raster_output", pptName = "ppt1")
}
