% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRI_from_accel_csv_v3b.R
\name{SRI_from_accel_csv}
\alias{SRI_from_accel_csv}
\title{Calculate Sleep Regularity Index from .csv Accelerometer Data}
\usage{
SRI_from_accel_csv(
  acceldir = c(),
  alloutdir = c(),
  dsdir = c(),
  col.timestamp = c(),
  col.accel = c(),
  outputdir = c(),
  rmc.col.acc = c(2:4),
  rmc.col.time = 1,
  rmc.nrow = c(),
  use.naps = TRUE,
  use.WASO = TRUE,
  use.miscal = TRUE,
  wr.SWV = TRUE,
  wr.raster = TRUE,
  ...
)
}
\arguments{
\item{acceldir}{Directory containing raw .csv accelerometer files}

\item{alloutdir}{General output directory}

\item{dsdir}{Directory for down-sampled files}

\item{col.timestamp}{Column of raw .csv files containing timestamp}

\item{col.accel}{Columns of raw .csv files containing x-y-z accelerometer data}

\item{outputdir}{Directory of GGIR output}

\item{rmc.col.acc}{Columns of accelerometer data in down-sampled files}

\item{rmc.col.time}{Column of timestamps in down-sampled files}

\item{rmc.nrow}{Number of accelerometer data rows}

\item{use.naps}{Specify whether 'naps' are included in analysis}

\item{use.WASO}{Specify whether Wake After Sleep Onset (WASO) periods are included in analysis}

\item{use.miscal}{Specify whether to filter out nights of data likely miscalculated by GGIR}

\item{wr.SWV}{Specify whether Sleep-Wake Vectors (SWV) are  output to file}

\item{wr.raster}{Specify whether sleep-wake raster plots are output to file}

\item{...}{}

\item{GGIRinstdir}{Directory of GGIR ver 2.0-0 files}
}
\value{

}
\description{
A wrapper that allows for direct calculation of Sleep Regularity Index (SRI) scores from accelerometer data. Runs functions
for down-sampling data, predicting sleep-wake timing (GGIR), accounting for broken sleep patterns,``
and calculating SRI.
}
\examples{
SRI_from_accel_csv(acceldir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/106_output/ds_output", col.timestamp = 5, col.accel = c(1:3))
}
