% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRI_from_accel_csv_v3b.R
\name{SRI_from_accel_csv}
\alias{SRI_from_accel_csv}
\title{Calculate Sleep Regularity Index from .csv Accelerometer Data}
\usage{
SRI_from_accel_csv(
  acceldir = c(),
  alloutdir = c(),
  dsdir = c(),
  col.timestamp = c(),
  col.accel = c(),
  outputdir = c(),
  rmc.col.acc = c(2:4),
  rmc.col.time = 1,
  rmc.nrow = c(),
  use.naps = TRUE,
  use.WASO = TRUE,
  use.miscal = TRUE,
  wr.SWV = TRUE,
  wr.raster = TRUE,
  ...
)
}
\arguments{
\item{acceldir}{Directory containing raw .csv accelerometer files}

\item{alloutdir}{General output directory}

\item{dsdir}{Directory for down-sampled accelerometer files}

\item{col.timestamp}{Column of raw .csv files containing timestamp}

\item{col.accel}{Columns of raw .csv files containing x-y-z accelerometer data e.g., c(1:3)}

\item{outputdir}{Directory for GGIR output}

\item{rmc.col.acc}{Columns of accelerometer data in down-sampled files}

\item{rmc.col.time}{Column of timestamps in down-sampled files}

\item{rmc.nrow}{Number of accelerometer data rows}

\item{use.naps}{Specify whether 'naps' are included in SRI calculation}

\item{use.WASO}{Specify whether 'wake after sleep onset' (WASO) periods are included in SRI calculation}

\item{use.miscal}{Specify whether to filter out nights of 'miscalculated' sleep onset/offset timing}

\item{wr.SWV}{Specify whether Sleep-Wake Vectors (SWV) are output to file}

\item{wr.raster}{Specify whether sleep-wake raster plots are output to file}

\item{...}{}
}
\value{

}
\description{
A wrapper that allows for direct calculation of Sleep Regularity Index (SRI) scores from accelerometer data (.csv format).

Runs three functions: (a) down-sampling data [ds_accel_csv], (b) predicting sleep-wake timing using GGIR [GGIR_from_csv],
and (c) calculating SRI after accounting for broken sleep patterns [SRI_from_GGIR].

Minimum required inputs: 'acceldir', 'col.timestamp', 'col.accel'.
}
\examples{
SRI_from_accel_csv(acceldir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/106_output/ds_output", col.timestamp = 5, col.accel = c(1:3))
}
