% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRI_from_accel_csv_v3b.R
\name{SRI_from_GGIR}
\alias{SRI_from_GGIR}
\title{Calculate Sleep Regularity Index (SRI) from GGIR Output}
\usage{
SRI_from_GGIR(
  outputdir = c(),
  alloutdir = c(),
  use.naps = TRUE,
  use.WASO = TRUE,
  use.miscal = TRUE,
  wr.SWV = TRUE,
  wr.raster = TRUE
)
}
\arguments{
\item{outputdir}{Directory of GGIR output}

\item{alloutdir}{General output directory}

\item{use.naps}{Specify whether 'naps' are included in SRI calculation}

\item{use.WASO}{Specify whether 'wake after sleep onset' (WASO) periods are included in SRI calculation}

\item{use.miscal}{Specify whether to filter out nights of data likely miscalculated by GGIR}

\item{wr.SWV}{Specify whether Sleep-Wake Vectors (SWV) are  output to file}

\item{wr.raster}{Specify whether sleep-wake raster plots are output to file}
}
\value{

}
\description{
Uses sleep windows and sustained inactivity bouts from GGIR output to calculate Sleep Regularity Index scores. Accounts for
multiphasic and broken sleep by identifying periods of 'wake' during GGIR-defined sleep windows and periods of 'napping' outside
GGIR-defined sleep windows. Uses sustained inactivity bouts to exclude days where sleep onset and offset times are likely
miscalculated. Runs across all "output_xxx" directories within 'outputdir', accounting for both multi-file and single-file GGIR output
structures.

Additional outputs: sleep-wake raster plots, sleep variables, summary of miscalculated nights, sleep-wake vectors.

Minimum required inputs: 'outputdir'
}
\examples{
SRI_from_GGIR(outputdir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/106_output/GGIR_output")
}
