% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SRI_from_accel_csv_v3b.R
\name{SRI_from_GGIR}
\alias{SRI_from_GGIR}
\title{Calculate Sleep Regularity Index (SRI) from GGIR Output}
\usage{
SRI_from_GGIR(
  outputdir = c(),
  alloutdir = c(),
  use.naps = TRUE,
  use.WASO = TRUE,
  use.miscal = TRUE,
  wr.SWV = TRUE,
  wr.raster = TRUE
)
}
\arguments{
\item{outputdir}{Directory of GGIR output - must be specified}

\item{alloutdir}{General output directory - specify if you wish}

\item{use.naps}{Specify whether 'naps' are included in analysis}

\item{use.WASO}{Specify whether Wake After Sleep Onset (WASO) periods are included in analysis}

\item{use.miscal}{Specify whether to filter out nights of data likely miscalculated by GGIR}

\item{wr.SWV}{Specify whether Sleep-Wake Vectors (SWV) are  output to file}

\item{wr.raster}{Specify whether sleep-wake raster plots are output to file}
}
\value{

}
\description{

}
\examples{
SRI_from_GGIR(outputdir = "C:/Users/dan_t/Documents/R/Biobank/SleepRegPackage/106_output/GGIR_output")
SRI_from_GGIR(outputdir = "C:/Users/dan_t/Documents/R/Biobank/from_alex/sampleresults5")
SRI_from_GGIR(outputdir = "C:/Users/dan_t/Documents/R/SRI_Study/Geneactiv_Data/sample/Geneactiv_five_bin_results_AusTz_15window")
}
