% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{SWS_from_binarySW}
\alias{SWS_from_binarySW}
\title{Extract Individual SWS Files from Binary Sleep-Wake Files}
\usage{
SWS_from_binarySW(binarySWdir = c(), tsCol = c(), binaryCol = c())
}
\arguments{
\item{binarySWdir}{Location of binary sleep-wake .csv files}

\item{tsCol}{Column containing timestamps}

\item{binaryCol}{Column containing binary sleep-wake vector}
}
\value{

}
\description{
Takes binary sleep wake files (containing a timestamp column and a binary sleep-wake column)
and converts to sleep-wake summary (SWS) format (file format used for input to SRI_from_binary function).

Minimum required input: 'binarySWdir', 'tsCol', 'binaryCol'
}
\examples{
SWS_from_binarySW(binarySWdir = "~/R/Biobank/binarySW", tsCol = 2, binaryCol = 1)
}
