% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepreg_allFunctions.R
\name{SWS_from_SWV}
\alias{SWS_from_SWV}
\title{Extract Individual SWS Files from single SWV File}
\usage{
SWS_from_SWV(
  SWVfile = c(),
  SWSdir = c(),
  use.naps = TRUE,
  use.WASO = TRUE,
  use.miscal = TRUE
)
}
\arguments{
\item{SWVfile}{Location of sleep-wake vector file}

\item{SWSdir}{Directory to write individual sleep-wake vector summary (SWS) files}

\item{use.naps}{Specify whether 'naps' are included in SRI calculation}

\item{use.WASO}{Specify whether 'wake after sleep onset' (WASO) periods are included in SRI calculation}

\item{use.miscal}{Specify whether to filter out nights of 'miscalculated' sleep onset/offset timing}
}
\value{

}
\description{
Takes SWV file (single file summary of sleep-wake, generated by 'SRI_from_GGIR')
and converts to individual sleep-wake summary (SWS) files. Specify whether output SWS
files account for naps, WASO, and miscalculated nights.

Minimum required input: SWVfile
}
\examples{
SWS_from_SWV(SWVfile="~/R/Biobank/Light_Data/127_evenfreq_output/SWV.csv")
SWS_from_SWV(SWVfile="~/R/Biobank/Light_Data/265_evenfreq_output/SWV.csv")
}
